[MAJOR] adding trellis speed, testing rembg, missing test.sh

+++ TRELLIS-TextoImagen3D/trellis/representations/mesh/utils_cube.py
@@ -32,7 +32,9 @@ def cubes_to_verts(num_verts, cubes, value, reduce='mean'):
         reduced[cubes[i][j]][k] += value[i][k]
     """
     M = value.shape[2] # number of channels
-    reduced = torch.zeros(num_verts, M, device=cubes.device)
+    # Ensure accumulator dtype matches value dtype to avoid scatter/scatter_reduce dtype mismatches under fp16
+    reduced = torch.zeros(num_verts, M, device=cubes.device, dtype=value.dtype)
+    # reduced = torch.zeros(num_verts, M, dtype=value.dtype)
     return torch.scatter_reduce(reduced, 0, 
         cubes.unsqueeze(-1).expand(-1, -1, M).flatten(0, 1), 
         value.flatten(0, 1), reduce=reduce, include_self=False)
@@ -49,7 +51,9 @@ def sparse_cube2verts(coords, feats, training=True):
 
 def get_dense_attrs(coords : torch.Tensor, feats : torch.Tensor, res : int, sdf_init=True):
     F = feats.shape[-1]
-    dense_attrs = torch.zeros([res] * 3 + [F], device=feats.device)
+    # Keep dtype aligned with feats (e.g., fp16) to avoid unnecessary casts and dtype mismatches
+    # dense_attrs = torch.zeros([res] * 3 + [F], dtype=feats.dtype)
+    dense_attrs = torch.zeros([res] * 3 + [F], device=feats.device, dtype=feats.dtype)
     if sdf_init:
         dense_attrs[..., 0] = 1 # initial outside sdf value
     dense_attrs[coords[:, 0], coords[:, 1], coords[:, 2], :] = feats
@@ -58,4 +62,3 @@ def get_dense_attrs(coords : torch.Tensor, feats : torch.Tensor, res : int, sdf_
 
 def get_defomed_verts(v_pos : torch.Tensor, deform : torch.Tensor, res):
     return v_pos / res - 0.5 + (1 - 1e-8) / (res * 2) * torch.tanh(deform)
-
